alert js추가

+++ src/main/java/com/soldesk/team_project/controller/AdminController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.SessionAttribute;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
@@ -20,6 +21,7 @@ import com.soldesk.team_project.dto.AdDTO;
 import com.soldesk.team_project.dto.LawyerDTO;
 import com.soldesk.team_project.dto.MemberDTO;
 import com.soldesk.team_project.dto.QuestionDTO;
+import com.soldesk.team_project.dto.UserMasterDTO;
 import com.soldesk.team_project.service.AdService;
 import com.soldesk.team_project.service.LawyerService;
 import com.soldesk.team_project.service.MemberService;
@@ -60,9 +62,15 @@ public class AdminController {
     // 일반 회원 관리
     @GetMapping("/memberManagement")
     public String memberList(
-        @RequestParam(value = "keyword",required = false) String keyword, Model model, 
-        @RequestParam(value = "searchType", required = false, defaultValue = "idx") String searchType) {
+        @RequestParam(value = "keyword",required = false) String keyword, Model model, RedirectAttributes redirectAttributes,
+        @RequestParam(value = "searchType", required = false, defaultValue = "idx") String searchType,
+        @SessionAttribute(value = "loginUser", required = false) UserMasterDTO loginUser) {
         
+        if (loginUser.getAdminIdx() == null) {
+            redirectAttributes.addFlashAttribute("alert", "관리자 권한이 필요합니다.");
+            return "redirect:/";
+        }
+
         List<MemberDTO> memberList;
 
         if (keyword == null || keyword.trim().isEmpty()) {
@@ -82,7 +90,13 @@ public class AdminController {
     public String memberSearch(
         @RequestParam("keyword") String keyword,
         @RequestParam("searchType") String searchType,
-        RedirectAttributes redirectAttributes) {
+        RedirectAttributes redirectAttributes,
+        @SessionAttribute(value = "loginUser", required = false) UserMasterDTO loginUser) {
+
+        if (loginUser.getAdminIdx() == null) {
+            redirectAttributes.addFlashAttribute("alert", "관리자 권한이 필요합니다.");
+            return "redirect:/";
+        }
 
         redirectAttributes.addAttribute("keyword", keyword);
         redirectAttributes.addAttribute("searchType", searchType);
@@ -92,8 +106,14 @@ public class AdminController {
     // 변호사 회원 관리
     @GetMapping("/lawyerManagement")
     public String lawyerList(
-        @RequestParam(value = "keyword",required = false) String keyword, Model model, 
-        @RequestParam(value = "searchType", required = false, defaultValue = "idx") String searchType) {
+        @RequestParam(value = "keyword",required = false) String keyword, Model model, RedirectAttributes redirectAttributes,
+        @RequestParam(value = "searchType", required = false, defaultValue = "idx") String searchType,
+        @SessionAttribute(value = "loginUser", required = false) UserMasterDTO loginUser) {
+
+        if (loginUser.getAdminIdx() == null) {
+            redirectAttributes.addFlashAttribute("alert", "관리자 권한이 필요합니다.");
+            return "redirect:/";
+        }
         
         List<LawyerDTO> lawyerList;
 
@@ -114,7 +134,13 @@ public class AdminController {
     public String lawyerSearch(
         @RequestParam("keyword") String keyword,
         @RequestParam("searchType") String searchType,
-        RedirectAttributes redirectAttributes) {
+        RedirectAttributes redirectAttributes,
+        @SessionAttribute(value = "loginUser", required = false) UserMasterDTO loginUser) {
+
+        if (loginUser.getAdminIdx() == null) {
+            redirectAttributes.addFlashAttribute("alert", "관리자 권한이 필요합니다.");
+            return "redirect:/";
+        }
 
         redirectAttributes.addAttribute("keyword", keyword);
         redirectAttributes.addAttribute("searchType", searchType);
@@ -124,9 +150,15 @@ public class AdminController {
     // 문의글 관리
     @GetMapping("/QnAManagement")
     public String questionList(
-        @RequestParam(value = "keyword",required = false) String keyword, Model model, 
-        @RequestParam(value = "searchType", required = false, defaultValue = "idx") String searchType) {
+        @RequestParam(value = "keyword",required = false) String keyword, Model model, RedirectAttributes redirectAttributes,
+        @RequestParam(value = "searchType", required = false, defaultValue = "idx") String searchType,
+        @SessionAttribute(value = "loginUser", required = false) UserMasterDTO loginUser) {
         
+        if (loginUser.getAdminIdx() == null) {
+            redirectAttributes.addFlashAttribute("alert", "관리자 권한이 필요합니다.");
+            return "redirect:/";
+        }
+
         List<QuestionDTO> newQuestions;
         List<QuestionDTO> completedQuestions;
 
@@ -150,7 +182,13 @@ public class AdminController {
     public String questionSearch(
         @RequestParam("keyword") String keyword,
         @RequestParam("searchType") String searchType,
-        RedirectAttributes redirectAttributes) {
+        RedirectAttributes redirectAttributes,
+        @SessionAttribute(value = "loginUser", required = false) UserMasterDTO loginUser) {
+
+        if (loginUser.getAdminIdx() == null) {
+            redirectAttributes.addFlashAttribute("alert", "관리자 권한이 필요합니다.");
+            return "redirect:/";
+        }
 
         redirectAttributes.addAttribute("keyword", keyword);
         redirectAttributes.addAttribute("searchType", searchType);
@@ -159,8 +197,14 @@ public class AdminController {
 
     // 광고 관리
     @GetMapping("/adManagement")
-    public String adList(Model model) {
+    public String adList(Model model, RedirectAttributes redirectAttributes,
+        @SessionAttribute(value = "loginUser", required = false) UserMasterDTO loginUser) {
         
+        if (loginUser.getAdminIdx() == null) {
+            redirectAttributes.addFlashAttribute("alert", "관리자 권한이 필요합니다.");
+            return "redirect:/";
+        }
+
         // 만료된 광고 비활성화
         adService.refreshActiveAds();
 
@@ -173,14 +217,28 @@ public class AdminController {
 
     // 광고 등록
     @GetMapping("/adRegistration")
-    public String registAdForm(@ModelAttribute("adRegistration")AdDTO adRegistration) {
+    public String registAdForm(RedirectAttributes redirectAttributes,
+        @ModelAttribute("adRegistration")AdDTO adRegistration, 
+        @SessionAttribute(value = "loginUser", required = false) UserMasterDTO loginUser) {
+
+        if (loginUser.getAdminIdx() == null) {
+            redirectAttributes.addFlashAttribute("alert", "관리자 권한이 필요합니다.");
+            return "redirect:/";
+        }
+
         return "admin/adRegistration";
     }
     @PostMapping("/adRegistration")
-    public String registAdSubmit(
+    public String registAdSubmit(RedirectAttributes redirectAttributes,
         @ModelAttribute("adRegistration")AdDTO adRegistration,
-        @RequestParam("imageFile") MultipartFile imageFile) {
+        @RequestParam("imageFile") MultipartFile imageFile,
+        @SessionAttribute(value = "loginUser", required = false) UserMasterDTO loginUser) {
         
+        if (loginUser.getAdminIdx() == null) {
+            redirectAttributes.addFlashAttribute("alert", "관리자 권한이 필요합니다.");
+            return "redirect:/";
+        }
+
         // 광고 이미지 업로드
         String desiredName = adRegistration.getAdImgPath();
 
@@ -211,7 +269,14 @@ public class AdminController {
 
     // 광고 상세
     @GetMapping("/adInfo")
-    public String showAd(@RequestParam("adIdx") Integer adIdx, Model model) {
+    public String showAd(@RequestParam("adIdx") Integer adIdx, Model model, RedirectAttributes redirectAttributes,
+        @SessionAttribute(value = "loginUser", required = false) UserMasterDTO loginUser) {
+
+        if (loginUser.getAdminIdx() == null) {
+            redirectAttributes.addFlashAttribute("alert", "관리자 권한이 필요합니다.");
+            return "redirect:/";
+        }
+
         AdDTO ad = adService.getAd(adIdx);
         model.addAttribute("ad", ad);
 
@@ -220,22 +285,37 @@ public class AdminController {
 
     // 광고 수정
     @GetMapping("/adModify")
-    public String modifyForm(@RequestParam("adIdx") Integer adIdx, Model model) {
+    public String modifyForm(@RequestParam("adIdx") Integer adIdx, Model model, RedirectAttributes redirectAttributes,
+        @SessionAttribute(value = "loginUser", required = false) UserMasterDTO loginUser) {
+
+        if (loginUser.getAdminIdx() == null) {
+            redirectAttributes.addFlashAttribute("alert", "관리자 권한이 필요합니다.");
+            return "redirect:/";
+        }
+            
         AdDTO modifyAd = adService.getAd(adIdx);
         model.addAttribute("modifyAd", modifyAd);
 
         return "admin/adModify";
     }
     @PostMapping("/adModify")
-    public String modifySubmit(@ModelAttribute("modifyAd") AdDTO modifyAd,@RequestParam(value = "imageFile", required = false) MultipartFile imageFile, Model model) {
-        if (imageFile != null && !imageFile.isEmpty()) {
-        String filename   = nowUuidName(imageFile.getOriginalFilename());
-        String objectPath = "ads/" + filename;
+    public String modifySubmit(@ModelAttribute("modifyAd") AdDTO modifyAd, RedirectAttributes redirectAttributes,
+        @RequestParam(value = "imageFile", required = false) MultipartFile imageFile, Model model,
+        @SessionAttribute(value = "loginUser", required = false) UserMasterDTO loginUser) {
 
-        var uploaded = storageService.upload(imageFile, objectPath);
-        
-        modifyAd.setAdImgPath(uploaded.url());   // 기존 경로 덮어쓰기
-    }
+        if (loginUser.getAdminIdx() == null) {
+            redirectAttributes.addFlashAttribute("alert", "관리자 권한이 필요합니다.");
+            return "redirect:/";
+        }
+
+        if (imageFile != null && !imageFile.isEmpty()) {
+            String filename   = nowUuidName(imageFile.getOriginalFilename());
+            String objectPath = "ads/" + filename;
+
+            var uploaded = storageService.upload(imageFile, objectPath);
+            
+            modifyAd.setAdImgPath(uploaded.url());   // 기존 경로 덮어쓰기
+        }
         
         adService.modifyProcess(modifyAd);
 
@@ -244,7 +324,14 @@ public class AdminController {
 
     // 광고 삭제
     @GetMapping("/adDelete")
-    public String deleteAd(@RequestParam("adIdx") int adIdx) {
+    public String deleteAd(@RequestParam("adIdx") int adIdx, RedirectAttributes redirectAttributes,
+        @SessionAttribute(value = "loginUser", required = false) UserMasterDTO loginUser) {
+
+        if (loginUser.getAdminIdx() == null) {
+            redirectAttributes.addFlashAttribute("alert", "관리자 권한이 필요합니다.");
+            return "redirect:/";
+        }
+
         adService.deleteProcess(adIdx);
 
         return "redirect:/admin/adManagement";
